Implement P04151 constexpr for <complex> (Again) (#397)
* <complex> Implement p0415r1 `constexpr For <complex> (Again)`

This implements p0415r1 `constexpr For <complex> (Again)` which can be found here https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/p0415r1.html

Unfortunately we cannot currently implement it fully because there are some issues with intrinsics:

nvrtc and msvc do not support those builtins that we need so we disable it for now. With msvc we can wor around by using bit_cast, but that requires massive code changes.

+++ libcxx/test/std/numerics/complex.number/complex.special/double_long_double_implicit.compile.fail.cpp
@@ -3,15 +3,16 @@
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+// SPDX-FileCopyrightText: Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES.
 //
 //===----------------------------------------------------------------------===//
 
 // <complex>
 
-// template<> class complex<float>
+// template<> class complex<double>
 // {
 // public:
-//     explicit constexpr complex(const complex<double>&);
+//     explicit constexpr complex(const complex<long double>&);
 // };
 
 #include <complex>
@@ -19,8 +20,8 @@
 
 int main(int, char**)
 {
-    const std::complex<double> cd(2.5, 3.5);
-    std::complex<float> cf = cd;
+    const std::complex<long double> cd(2.5, 3.5);
+    std::complex<double> cf = cd;
     assert(cf.real() == cd.real());
     assert(cf.imag() == cd.imag());
